server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
torch==2.0.1
transformers==4.31.0
numpy==1.24.3
scikit-learn==1.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
# Core
flask==2.3.3
flask-cors==4.0.0

# ML/NLP
torch==2.0.1
transformers==4.31.0
scikit-learn==1.3.0
numpy==1.24.3
scipy==1.10.1

# Voice Processing
SpeechRecognition==3.10.0
librosa==0.10.0
soundfile==0.12.1

# Audio
pydub==0.25.1

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0
